Kansas munging improvements
