+++ src/ini_files_for_results/Kansas/ks_20g_ks_house_official.ini
@@ -1,6 +1,6 @@
 [election_results]
 results_file=Kansas/2020_General_Election_Kansas_House_of_Representatives_results_by_precinct.xlsx
-munger_list=ks_gen_main,ks_gen_johnson_count_from_B,ks_gen_shawnee_count_from_B,ks_gen_sedgwick,ks_gen_wyandotte_4_line_header_first_count_col_3
+munger_list=ks_gen_main,ks_gen_johnson_count_from_B,ks_gen_shawnee_count_from_B,ks_gen_sedgwick,ks_gen_wyandotte_4_line_header_first_count_col_3,ks_gen_wyandotte_3_line_header_first_count_col_3,ks_gen_wyandotte_4_line_header_first_count_col_3_merged_rows
 jurisdiction=Kansas
 election=2020 General
 results_short_name=ks_20g_kshouse
