+++ src/mungers/ks_gen_shawnee.munger
@@ -62,7 +62,7 @@ Party={<count_header_2>,^(\w\w\w) .*$}
 # Values to ignore (optional) #
 [ignore]
 ## E.g: Candidate=Total Votes Cast,Registered Voters ##
-Candidate=Write-in Totals
+Candidate=Write-in Totals,Write-in
 
 # Lookup formula sections #
 ## Required when foreign keys are used in munge formulas and    ##
