+++ src/mungers/ks_gen_wyandotte_3_line_header_first_count_col_3.munger
@@ -4,56 +4,63 @@
 #### File type must be one of: excel,json-nested,xml,flat_text
 file_type=excel
 #### Counts are found in one way of: by_name,by_number
-count_location=by_number:1,2,3,4,5,6,7,8,9,10,11
+count_location=by_number:3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24
 
-# Optional parameter for excel or flat_text
-#rows_to_skip=
+merged_cells=yes
 
+################################################
 ## Sometimes required format parameters:
 #### for flat_text file type:
-#flat_text_delimiter=
+flat_text_delimiter=
 #### if count_columns_specified is 'by_name':
-#count_fields_by_name=
+count_fields_by_name=
 #### if count_columns_specified is 'by_number':
 #### if 'in_count_headers' is in munge_strings
 #### (start numbering from first unskipped row):
 count_header_row_numbers=0,1,2
-#### if 'constant_over_file' is in munge_strings:
+#### if 'constant_over_file' is in munge_strings (NB: give value for each in .ini file):
 constant_over_file=CountItemType
 #### if file type is flat_text or excel and count_columns_specified is 'by_name'
 #### (start numbering from first unskipped row):
-#count_field_name_row=
+count_field_name_row=
 #### if file type is flat_text or excel and not all rows are data:
 #### (start numbering from first unskipped row):
 noncount_header_row=0
 
-
+################################################
 ## Optional format parameters:
 #### for any file type:
-#thousands_separator=
-#encoding=
+thousands_separator=,
+encoding=
+
 #### for a flat_text or excel file type:
+###### if field names are not given in file
 #all_rows=data
+###### if there are multiple blocks of data per page, each with its own headers
+multi_block=yes
+
 #### for excel file type:
-sheets_to_read_names=Johnson,JOHNSON
+sheets_to_read_names=Wyandotte,WYANDOTTE
 sheets_to_read_numbers=
 sheets_to_skip_names=
+
 #### for xml file type
 nesting_tags=
 
 # Munge formula sections (required if in munge_strings list) #
 [munge formulas]
-ReportingUnit=<sheet_name>;<PRECINCT NAME>
-Candidate=<count_header_2>
+ReportingUnit=<sheet_name>;<column_1>
+Candidate={<count_header_2>,^(?:\w\w\w |)(.*)$}
 CandidateContest=<count_header_0>
-Party=<count_header_2>
+Party={<count_header_2>,^(\w\w\w) .*$}
+
 
 
 # Values to ignore (optional) #
 [ignore]
 ## E.g: Candidate=Total Votes Cast,Registered Voters ##
-ReportingUnit=JOHNSON;COUNTY TOTALS,Johnson;COUNTY TOTALS
-Candidate=Write-in,Under Votes,Over Votes
+ReportingUnit=WYANDOTTE;Totals,Wyandotte;Totals,WYANDOTTE;
+Candidate=Write-in Totals
 
 # Lookup formula sections #
 ## Required when foreign keys are used in munge formulas and    ##
