+++ src/reference_results/Kansas.tsv
@@ -2,7 +2,7 @@ Jurisdiction	Election	Contest	ReportingUnit	VoteType	Count	Status	Source	Note
 Kansas	2020 General	US President (KS)	Kansas	total	1457960	official-final		
 Kansas	2020 General	US House KS District 3	Kansas	total	410418	official-final		
 Kansas	2020 General	KS Senate District 15	Kansas	total	23043	official-final		
-Kansas	2020 General	KS Senate District 6	Kansas	total	22538	official-final
+Kansas	2020 General	KS Senate District 6	Kansas	total	22583	official-final
 Kansas	2020 General	KS Senate District 35	Kansas	total	28779	official-final
 Kansas	2020 General	KS House District 2	Kansas	total	10874	official-final
 Kansas	2020 General	KS House District 32	Kansas	total	3780	official-final
